backend/fastapi_ml/requirements.txt
fastapi
uvicorn
pandas
numpy
scikit-learn
joblib
python-dotenv
PyPDF2
requests
pydantic